requirements.txt
ipython==8.4.0
keras==2.9.0
numpy==1.22.3
pandas==1.4.2
Pillow==9.2.0
streamlit==1.11.1
tensorflow-cpu==2.9.1

